figuring out formatting and color pallette after resolving NPM package issues
