+++ client/src/App.css
@@ -1,4 +1,4 @@
-.App {
+/* '.App {
   text-align: center;
 }
 
@@ -35,4 +35,11 @@
   to {
     transform: rotate(360deg);
   }
+} */
+
+h1 {
+  color: rgb(95, 201, 195);
+  font-family: "Lucida Sans", "Lucida Sans Regular", "Lucida Grande",
+    "Lucida Sans Unicode", Geneva, Verdana, sans-serif;
+  font-size: xx-large;
 }
