+++ package.json
@@ -13,7 +13,6 @@
   "license": "ISC",
   "homepage": "",
   "dependencies": {
-    "@mui/material": "^5.2.0",
     "express": "^4.17.1",
     "if-env": "^1.0.4",
     "mongoose": "^5.11.17"
