+++ server/package.json
@@ -11,7 +11,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "apollo-server-express": "^2.19.0",
     "express": "^4.17.1",
     "mongoose": "^5.11.6"
   },
